Fixes issue 32237: Payment In and Payment Out lines are not exported to CSV.
The problem was that export to CSV does not work properly in some subtabs without
link to parent column (e.g.Lines subtab of Payment In/Out). Before, It was used
"tabId" to obtaing data for  export to CSV and the value of the "tabId" was sent
in a variable called "tab". The Criteria that is run when it is displayed the tab
uses a property named "tabId" instead of "tab".

The solution for this case (and other related cases) is to adapt the code that
exports to CSV to use the same Criteria run when displaying the tab. Now export to
CSV uses tabId property. Besides it has been refactored some piece of code related
with export data in DataSourceServlet to use "tabId" instead of "tab".