[pi-security-hqlinjection] Code changes to fix Action regarding selector
The SelectorDataSourceFilter class was not working properly, it was only
taking into account the whereClauses when  the selector had a filter expression.
Every selector does not have a filter expression, so, it was not working in that cases.
Now this is taken into account and when the selector has no filter expression, the
hql where clause is taken.