[pi-security-hqlinjection] Preference added
The code which handles the allow where parameter preference has been added.
Now, if it comes a _where parameter and the preference is set to Y, this where
clause will be taken into account. If a _where parameter comes but the preference
is N or it is undefined, a exception is thrown.