Fixed issue 32151: It wasnt possible configure text referenced cols. properly.
It was not possible to configure the Allow Filtering property in colums that used Text reference.
The problem was in the UIDefinition.removeAttributeFromString.

Here the canSort and canFilter properties were being deleted and that was not correct, because in the case of
the "help" column of the Module window, for example, when the method was invoked, it was only desired to remove
the "canSort" property and not the canFilter.
The method has been changed, now, the passed String argument is the one which will be deleted instead of deleting
canSort or canFilter directly.