Fixes issue 32396: cannot filter grid when it is grouped by a column.
Having a column grouped, if a filter was applied in a grid, it was not
possible to apply it.A js error was thrown:
Uncaught TypeError: this.data.willFetchData is not a function.

The problem was in:"this.data.willFetchData(....))" statement which
incorrectly assumed this.data to be an object and willFetchData a function
within it. This is not true when in this situation where this.data is a
Tree class.

It has been fixed by taking into account this case by checking that "this.data" defines the
function.