fixed issue 32152: Some colums were modified even with no modules in devel.
The AllowSorting and AllowFiltering columns were modified even with no module in development.
That behavior is not correct and the AD_COLUMN_MOD_TRG.xml trigger has been modified to handle
this situation. Now if there is no module in development and the old value of these columns is
not the same as the new one (if it has been modified) an error is shown.