Fixes issue 32437: Wrong orders created by the Copy Record process
Fixed the wrong Tax Amount and Total Gross Amount calculation in orders created by the Copy Record process
if price list includes taxes. These amounts are calculated when related triggers are fired, due that when
the order is cloned from the original order was necessary to reset calculated amounts and then triggers updates
them again when lines are added to the cloned order.