[pi-security-hqlinjection] An informative comment added
An informative comment about the WHERE_AND_FILTER_CLAUSE constant has been added.
It was not intuitive why the constant was cleaned up. This is done because this is
the constant that will contain the where and filter clause and it must be empty at this
poin because the where and filter clauses will be set and stored there after.