[pi-security-hqlinjection] The warn message was shown too many times
In the Selectors, the warn message in the log was shown two times in the case of having _where and the
preference set to "Yes". This was like this because inside a private method of the class SelectorDataSourceFilter
the preference's value is checked but the warning was thrown and that was not necessary at that point because
it has been thrown previously also.