[pi-security-hqlinjection] Fixed some incorrect behaviors at the boundaries
In the DataSourceServlet class, if the "filterClass" variable was hardcoded to null,
the selector filter was never done and the selectors showed too many registers.

In the DefaultDataSourceService class, in the addFetchParameters method,  there was
a condition which checked if it was a selector or not and if it was not a selector, all
the logic which handles the whereClause was done. The problem was that it was possible to
send a selector parameter to jump over all the functionality and inject any wanted where.

These two cases have been fixed by creating a new parameter called "WHERE_CLAUSE_HAS_BEEN_CHECKED".
This parameter is first initialized to false in the "DataSourceServlet" and after that, if
the where is set or just checked in the "SelectorDataSourceFilter" class, it is set to true.
After, in the DefaultDataSourceService, a condition has been added. If this parameter has the
value set to "true", then it means that it is a Selector and the where clause has been set and
checked in the "SelectorDataSourceFilter", so it is not neccessary to do all the chekcs again.

This fixes the two incorrect behaviors because now, if the "filterClass" variable is set to null,
the logic of the "DefaultDataSourceService" is executed.