fixes issue 32474: [object Object] appears after clearing value in selector
Having singleDateMode as false after clearing the filter content of the date causes the return of an empty object in the getValuesAsCriteria method of the grid. The this.singleDateMode was being set to true in clearFilterValues function, but this statement was deleted to solve issue 31705, because it was preventing to show the range dialog after clearing the filters, when pressing the Enter key.

To solve both problems the change regarding setting singleDateMode in clearFilterValues has been reverted, and in order to ensure that the range dialog is displayed when it should be, a new check has been included in the keyPress method.