Fixes issue 32460: Improves log. Supports true, yes and on literals
If modulescripts are skipped, it is shown in hte log. Also now it is possible to use 'yes', 'true' and 'on' as the property value to specify that the modulescripts should be exec
uted, as it is defined in the standard (see [1]).

Also the execute.modulescripts property is now defined only once in the build.xml.

[1] https://ant.apache.org/manual/develop.html

--HG--
extra : rebase_source : d8358e7345dfe286aefd2b318accc8e336c2e1eb