[pi-security-hqlinjection] Fixed the whereParameterIsNotBlank method
This method was not taking into account the fact of having a _where parameter
null but as String. So a condition has been added. Now it takes into account not
only that the _where parameter is not null, blank or an empty string, now it checks
also that is not the "null" string.