Fixes issue 32392: Wrong value in qtyreserved column of m_storage_pending in sales quotation.
When the "Create Order process" was launched from a Sales Quotation that generates pending storage, the ordered quantity in the M_STORAGE_PENDING was the double than the original for a recently created product. It was caused because a double call of the M_UPDATE_STORAGE_PENDING PL. Inside the C_ORDER_POST PL (it creates the entry inside the M_STORAGE_PENDING with the quantity reserved), and then the lines were processed again, the M_UPDATE_STORAGE_PENDING was called again and it updates the previously created entry increasing the quantity reserved once more time.
The process is working as expected now.

--HG--
extra : rebase_source : 90562be61c2f22eb79db61e6e0d233c9592347d6