Fixes Issue 32424:New Account not created for Business Partner
ORDER BY 1 DESC clause is now changed to DEFAULT that is ascending
order of the level of organization. As unique entry for every
acctschema is allowed. In the descending order case top most organization
setting were used to create entry in accounting tab. In the case mentioned
in the issue, F&B International Group org has level 3 and was first
to be inserted for schema F& International Group USD Account and for this
configuration create new bp account setting did not exists so deafult account
entry was made and though F&B US Inc had setting of new bp account but the
same acctschema could not be repeated in this case.