Fixed issue 32600: Export csv was not working fine
The csv files where not correctly exported from the client if the client was in a different
timezone.
The problem was that in the AdvancedQueryBuilder class, the way of setting the date time was
not correct.
The Calendar.HOUR constant was being used instead of Calendar.HOUR_OF_DAY. The first, is to
set or get the hour indicating the morning or the afternoon, for example the 11:00:00 PM, the
hour is 11. The second one, the one that is used now, is used for the 24 hours clock, so, if
the time isn 11:00:00 the HOUR_OF_DAY is 23.