Fixed issue 32819:An error is thrown when a Query List Widget is fetched
The problem was that an exception is thrown because entity OBKMO_WidgetClass
is not readable by the user 100. The problem occurs when it is try to get a
WidgetClass to check security:
OBDal.getInstance().get(WidgetClass.class,parameters.get("widgetId"));

It has been resolved by adding a OBContext.setAdminMode(true) before obtain
WidgetClass class.