Fixed issue 32795: Calculate permissions for some target entities selectors
The problem was that it is not take into account access permissions for
some entities selectors. This selectors are in tabs of P&E WindowReferences.

It have been implemented new logic in EntityAccessChecker to take into
account target entities of selectors allocated into tabs of WindowReferences.
For example in Payment In > Add Details > GL Items section > G/L Item selector.