Fixed issue 32831: Consistency: Split discounts with non grouped products
The discount engine calculates the promotions of the receipt merging all
product, without taking in consideration the check of the Grouped
product.

The solution is to don't merge the promotions of non grouped products and
calculate all promotions to apply in the receipt checking if the involved
products are non grouped products.

--HG--
extra : rebase_source : 1427fa9a06e2db665380161c85a8a8f5a6d50b51