Fixed issue 32833: Tree view is not working properly
When a node in tree view is moved, entity used to check
security access is null. This exception is shown:
OBSecurityException - Entity null is not readable by the user 100
The problem was that getEntity() method can't retrieves an entity.

Now it have been resolved by getting entity from table
id (referencedTableId).