Fixes issue 31912: It is not possible to delete a product with a permanent cost
When a product is removed, the M_COSTING trigger is fired in cascade to delete the related costs, and it don't allow to remove the existing permanents costings associated to the product. To allow to remove the product with permanent costs was modified the M_COSTING_TRG to remove the permanent costs if the product have not related documents.