Fixed issue 32912: If a child tab is opened an error was got after refreshing
This was caused because having an empty string as where clause was not avoided,
so, in the code there was a line which was adding an "and" after a where clause
and this was also happening when the where clause was an empty string.
To fix this, the condition has been changed, now it checks if it is an empty string
instead a null where clause.

--HG--
extra : rebase_source : 8e413bab3f901c9061c2e7209cfc51cbf8735462