Related to issue 32912: Test added to check the correct behavior when refresh
When the WHERE_AND_FILTER_CLAUSE parameter was an empty string, the
behavior was wrong, it was adding an "and" at the beginning of the
where clause and this was not corret. This behavior has been fixed and this
test has been added to check that the datasource response is now correct.
The parameters used in the datasource of this test are the same paramertes
that where used in the datasource of the child tab of the "Price List" window
after refreshing the browser.
This is to imitate the behavior of refreshing the browser having a child tab opened.
The fact of having these parameters, makes that the WHERE_AND_FILTER_CLAUSE parameter
is the empty string, so we can now check that the request works fine now.