Fixed issue 32950.Use HashSet instead of ArrayList to avoid duplicates
The product ids are now stored using HashSet as this implementation does not
allow duplicates.

The method to create the vars in some scenarios is throwing a OBException that
is properly catch. It has been changed the try to avoid the OBException and
remove the error message from the logs.