Fixed issue 32923: There was a NPE in the SelectorFieldPropertyCallout
There was a NullpointerException in the SelectorFieldPropertyCallout because
in the printPage method, there was not being taken into account the fact of
having a property of a computed column and the entity was not correctly retrieved.
This have been fixed checking if it is a computed column and generating the entity
correctly if so.

--HG--
extra : rebase_source : ee7be7f6a4f7e1882c6b5753ada00fcbb47e57ed