Fixed issue 32704: Base test for weld has conflicts with analytics module.
The problem was when a test that uses WeldBaseTest is thrown in an enviroment
with analytics module installed. It is a known issue in Jersey Servlet.

Setting this property to "true" is avoided throwing the error. It is a workaround
implemented in jersey-servlet-1.19.jar (CDIExtension.class) for Weld proxy bug.

--HG--
extra : rebase_source : f04c9dfeea27d2e1f4e5a5c4b8b7b47886d09c1f