related to issue 33015: prevent warning on DefaultJsonDataService
After reopening the selector pop-up having a selected record, a bad request was being identified by the DefaultJsonDataService: the request included the _targetRecordId parameter but the _directNavigation parameter was not present.

This was happening because the selection information was not being cleared by the changes introduced on the openSelectorWindow function. Before [1], this cleaning was done by the invalidateCache method. Now we are cleaning the selection by explicitly calling to deselectAllRecords().

[1] http://code.openbravo.com/erp/devel/pi/rev/9105dfe33d79e5b7ba39a2680064f7c48c402701