Fixed issue 33051: The "hasGeneralAccess" method is called twice
The HttpSecureAppServlet.hasGeneralAccess method is called twice
when a window is refresed or opened, for example "Sales Order".
This is not rigth because it executes more queries than needed.
To fix this a piece of code have been refactored in the "DataSourceServlet"
class.
In the "doPost" method, there is a call to the "hasAccess" method, this method
is the one which calls the "hasGeneralAccess" and this is the one that is called twice.
The problem was that inside the "doPost" method exists also a call to the "doFetch" method,
this method calls the "hasAccess" method and this calls the "hasGeneralAccess" method.
So, when the doPost method called the fetch method, the "hasGeneralAcces" method was executed
twice because of the call inside the "doPost" and the call inside the "doFetch" method.

This has been fixed by moving the code which was doing the call to "hasAccess" inside
the "doPost" method, now it has been placed after the call to the "doFetch".

--HG--
extra : rebase_source : 2b73225f45cabeea8229935758dfad5bdace3070