Fixes issue 33043: Permissions problems after process payment solved
Role defined for an specific org is not able to see data from sister orgs.

Regression was introduced by the commit associated to issue 28931:
http://code.openbravo.com/erp/devel/pi/rev/78fd9bade67beea0399977fcdc05fd267c4af8c4

Changes were rolled back.
Also this patch includes fix for the issue 28931, avoiding to remove from the Writable Organizations List of the OBContext object, the Organization of the sequence of the payment's document type, if this organization is already in the list of accesible organizations of the current logged role.

--HG--
extra : rebase_source : 0fe90d843e2af8dc9d2451231acf9a0642b3301d