Fixed issue 33073: It was not possible to use _where param in JSON webservices
It was not possible to filter using the "_where" parameter. The "whereAndFilterClause"
parameter was needed instead.
This was not correct because in the Webservices we must mantain the use of the "_where"
parameter to filter the data.
This has been fixed in the "JSONRestServlet" class. Before calling the fetch method, the
"whereAndFilterClause" has been overwritten with the value of the "_where" parameter. Now
when a user uses "_where" in a query of a JSON webService, that value will be set in the
"whereAndFilterClause" parameter and will work properly.

--HG--
extra : rebase_source : 2f38c6bc21e40f4d3966df3f4f0bd1cd9deb410d