Fixed issue 33064: form widgets cant be added to workspace
The problem was related with widgets that can be embedded
into a form. In some cases, these widgets should not be
placed into the workspace.

To resolve this problem in a generic way, it has been
implemented a flag (Available_in_workspace) in the widget
definition that determines when a widget is available to
be shown in the workspace.

Now, any widget that does not have selected the new flag,
will not appear in the list of widgets to place into the
workspace.