Fixed issue 32882:  Add products and create new receipts quickly induces severe failures in Web POS
The proble is that the calculateReceipt function gets stuck in the execution if
a new ticket is created in the middle of the process.

The solution is to disable also the new and delete button from the point of sale
window during the execution if the calculateReceipt fucntion.

To achieve this, one new synchronizationHelper has been created
(calculateReceipt) and two of them have been deleted (applyPromotions and
calculateGross).