fixes issue 33208: Direct link to documents in the workspace is not working
The problem happened having a saved view with a date field as the first one of the grid. In this case the date item was firing an extra datasource call causing the bad behavior. This extra call was being performed because the date field was detecting wrongly the deletion of its content. Now an extra check is added: the empty filter is applied if the newValue is '' and also if the grid was being filtered by this field.