fixes issue 33321: Implement isSameTab for Process Definition based windows
Process definition based windows can be opened in more than one tab at the same time. This allows, for example, to see the parameters used to run the process in case the process window is reopened after the execution.

This is the current behavior of such windows. We are just implementing the isSameTab method to remark this behavior.

--HG--
extra : rebase_source : e792ade26f2609344996a1c06e485a26b0f7f606