Fixed issue 32495: The "Create Lines From" process is fixed.
The "Create Lines From" process was creating multiple product lines when the
OK button was pressed several times. This could be reproducuble in environments
with slow internet connections in those environments the OK button could be clicked
a lot of times. This was not correct, the OK button should be disabled after pressing
it the first time.
The "submitThisPage" function of the "CreateFrom_Invoice.html" has been changed. Now,
in this function the buttons clicked are disabled and thanks to this, the OK button is
unable to be clicked again after the first time.

--HG--
extra : rebase_source : 84123b54fddf1a39e3287b719af355e838064da4