Fixed issue 33003: The titles in the CSV were wrong
If there was a property field, the titles of those columns were wrong
when exporting to csv.
The problem was in the QueryJSONWriterToCSV inner class of the DataSourceServlet class.
Here wher exporting to csv there was only taking into account if the column of the field
was the same of the column checked in the iteration. The problem occurs when there are
two or more fields which have the same column.
Now the condition has been changed, first it checks if the current field is a property field
and if the property is the one we really want.

--HG--
extra : rebase_source : 36d8754e43314bd4efbdb8c9f85ce6aa8a01b409