Fixed issue 33216: The onChange Event didn't work properly
The problem was that the client side callouts were not working properly the second time a record of a pick and edit window was modified in grid view.
The expected behavior is that a value put in one concrete field is copied to another when the first one looses the focus.
The problem was when the productCharacteristics item was present.
In the ob-pick-and-execute-view.js there was a condition wich was setting the
property "readOnlyEditorType" to "OBTextItem" for the fields including the productCharacteristics one.
This was causing the problem afterwards when the check "his.getEditorType(editField, values) != this._editRowForm.getEditorType(liveItem)"
was done in the ListGrid.js because the left part was retrieving "OBTextItem" instead of "OBCharacteristicsGridItem".
This has been fixed by creating this property for the characteristics items. To do so, CharacteristicsGridItem.java has now
a new method which retrieves the value "OBCharacteristicsGridItem". In the ftl now, the property "readOnlyEditorType" is set.
Now the condition of the ob-pick-and-execute-view.js has changed and it checks if there exists the "readOnlyEditorType" property
for the current field. If it exists, it will not do anithing and if not, the property will be set as "OBTextItem" as it was
doing before. Now, the characteristic items will have that property set as "OBCharacteristicsGridItem" and it will work properly.

--HG--
extra : rebase_source : a9666858ca12665bf2f4ba48523a428ab4862dd0