Fixed issue 32671: Display logic based on a prop field was not working fine
A field with a display logic based on a property field was not working fine.
In form view, when saving a record, the value of the display logic based on
a property field was not being applied. But if the record was closed and opened
again the values were being correctly shown. This was because when saving the
record clicking the button, the form was not being redrawed.
To fix this the ob-view-form.js has been changed.
Now, when clicking the save button, the form is redrawed and the value of the
field with the display logic is updated.

--HG--
extra : rebase_source : bcb60defd1a5643e5b8fd736b1cf5ebcbd8dbcae