Fixed issue 32260: Editing records in grid view was not working properly.
Editing in grid view was not working properly when all the fields of the record were readonly.
The problem was in "ob-view-grid.js" in the "startEditing" function. In this function, the focused field is set, but when every field is readonly there was not setting the value, so it was undefined. This was causing the failure.

To fix this a new condition has been added. Now, it is checked if the colNum is undefined and if the new function "isThereAnyEditableField" returns false. If these conditions are met, it means that the focus is not set yet and that there aren't any editable fields. Because if there were any editable fields, not setting the focus at that point would not be a problem, smartclient puts the focus there.

--HG--
extra : rebase_source : 9188356e6c1e050db3630c962ac7f406f095f086