Fixes issue 33447:The List of modules included in OB3 distribution is available
The list of modules included in the OB3 distribution are now available using the ModulesInOB3Distribution.getModules() method. The list has been extracted from the HelpWindow cla
ss, which now references the extracted listFixes issue 33447:The List of modules included in OB3 distribution is available

The list of modules included in the OB3 distribution are now available using the ModulesInOB3Distribution.getModules() method. The list has been extracted from the HelpWindow cla
ss, which now references the extracted list.