Fixes issue 33448: DatasetServiceTest checks only modules in OB3 distribution
The DatasetServiceTest class takes for granted that some rules applies to all datasets (i.e. that the dataset table where clause can be executed as is), but that cannot be met in some modules outside the OB3 distribution.

Now the test only checks the datasets defined in the modules included in the OB3 distribution.