Fixed issue 33478: Implicit filters were not working fine on updated instances
When a grid configuration was done in one instance and after that the instance was updated, the implicit filters were not working fine. The blue popup expressing that a filter is applied appeared but there was no filter applied.
The problem was in the isFilterClauseApplied function of ob-grid.js. There only was taking into account boolean values. Now, it takes into account any value.

--HG--
extra : rebase_source : e13adcc0fe60831bf3896de7c2ff62848caa488d