Fixed issue 33172, Fixed issue 32875: CashUp changes
The CashUp must be the last model to be synchronized in the backend. Sometimes,
this is not the case, and some synchronization models (such as Order and
CashMgmt) are not related with any cashup.

The solution, is to send the cashUp report within the synchronization models
(Order, CashMgmt, VoidLayaway and the CashUp itself).