Fixed issue 33507: Incorrect Audit info was generated sometimes.
There was a problem when creating audit info data by some sqlc processes, sometimes wrong data was being created. The problem was that when the processes used the getTransactionConnection, the context information table was not correctly set. This table is set by the SessionInfo.setDBSessionInfo method.
To fix this, a call to this method has been added in the ConnectionProviderImpl.getTransactionConnection() and in the DalConnectionProvider.getTransactionConnection().

--HG--
extra : rebase_source : 79444842d611664a4390182ed88499fbd2e0b962