fixes issue 33476: Wrong behavior in created/updated by filter with saved view
After applying the saved view, regarding the 'Created By' field, the displayField property ('_identifier') of the filter editor was being replaced with the display field of the field itself ('createdBy'). Therefore, after applying the saved view the values being displayed within the filter drop-down were the 'Created By' values for the entity referenced by the 'Created By' field (User). The same applies for the 'Updated By' field.

The root cause of this problem was that the fkField property for 'Created By' and 'Updated By' fields was not set to true. This allows to configure properly the filterEditorProperties on the client side for those fields, including the displayField whose value is set to '_identifier'.

From now the displayField property is set for the 'Created By' and 'Updated By' on the prepareGridFields() function, this way it will not be overridden after applying the saved view.