fixes issue 33572: Ability to define a Report Template as default
A new field is added to the Report Template tab in the Document Type window. It is a checkbox that can be used to mark a template as default. If a template is marked as default, is the one that will be selected when printing the document. Also a new trigger has been created in order to allow just one default template per document type.

--HG--
extra : rebase_source : 0b9f14bf2877be615a01e11b53bf6d14babe7df2