fixes issue 33577: Not possible to translate tooltip message of the date item
The pickerIconPrompt property is now defined for OBDateItem. This was already defined and working fine for the date item displayed in the filters. But the definition of the pickerIconPrompt, fromDateOnlyPrefix and toDateOnlyPrefix was being set twice. Now those properties are set just in the ob-form-styles.js file

--HG--
extra : rebase_source : 4bf67bed2bbaf9383e96d9f2b32a128b70c449f3