Fixes issue 33489: Can't post CAD if period was created previous to 3.0PR14Q4
Cost Adjustments feature was included in Openbravo 3.0PR14Q4, so periods created in lower versions doesn't includes the Cost Adjustment Document. When the instance is upgrated, these periods was not updated and it causes that CAD can't be posted. To include this document to all these periods was updated the CreateAccountingConfiguration module script to take into account control periods that has not included the CAD. Also was updated the Module Script Execution Limits to the 29718 version.