fixes issue 33669: Extra DS call after opening window with summaries on subtabs
An unneeded datasource request was being fired after opening a window having summary functions on a child tab. This was caused because the getSummaryRow() is invoked when drawing the window components. If the summaryRow is already generated at this point, the summaries are recalculated by Smartclient.
In this case, the request is not needed, because the summaries in the child tab will never have any value as no record is selected on the parent.
To prevent this extra request, the grid now checks in the getSummaryRow() whether it is located in a child view and in this case, if the parent grid is being loaded, then the request will not be fired.