Fixes issue 33666: Zero amount instead of withdrawal amount in Add Payment
When creating a payment out with Add Payment process in Financial Account window (from Transaction tab or Match Statement | Add Transaction process), actual payment will be loaded with withdrawal amount.
If transaction withdrawal amount is different than 0, actual payment will be loaded with withdrawal amount but will not be distributed into order/invoices in the grid. If total amount is lower than withdrawal amount, actual payment will not be updated with total amount and we will be able to leave the difference between withdrawal amount and total amount as credit. If total amount is higher than withdrawal amount, actual payment will be updated with total amount and we will be able to pay more amount than withdrawal amount.
If transaction withdrawal amount is 0, actual payment will always be updated with total amount and it will not be possible to generate credit in this payment.

We have reverted changes done in ob-aprm-addPayment.js for issues 33465 and 31392, and we will retrieve bslamount as deposit amount - withdrawal amount also in Transaction tab, as it is done in Add Transaction process.
With this, opening Add Payment from both Transaction tab and Add Transaction process, will have the same behavior as it used to when opening from Add Transaction process in 16Q1 (before fixing issue 31392).

--HG--
extra : rebase_source : f0420e48ba7f5ed1ff68b9b5d0c8621ef2bfef1d