Fixed issue 33639: Update Characteristic was not showing the correct data
If there was a product with a subset of a product characteristic, the update characteristic process was not showing the correct data. Instead of showing the products of the subset of the product characteristic, it was showing every product of the product characteristic.
To fix this, the where clause of the selector reference has been changed, making it more restrictive in the case of having a subset.

--HG--
extra : rebase_source : 19dad6d4bd2ed8b136182e52faed468ef895e783