Fixes issue 33599: Can't validate a costing rule after one with FIFO algorithm
Can not validate a new Costing Rule after validating onw with FIFO algorithm if there exists negative stock.
Problem is that FIFO Costing Algorithm is not able to calculate the cost of the closing inventory transaction with negative stock, because there will be more quantity in the Output Stack than the available in the Input Stack.
In that case, "NoCostCalculated" error message will be shown when validating the new Costing Rule until negative stock is regularized.