[Stock Screen] Execute Store Stock and Other Store Stock callbacks even the product left subwindow is not shown yet
This change has been done because a new hook is being executed after beforeSetShowing method in product left subwindow. As the execution of the hooks could be asynchronous, now the product popup is opened in the callback of the hooks in order to do it in order. So, now the popup is shown after the Other Store Stock and Store Store callbacks are executed, so those two checks must be removed.

--HG--
extra : transplant_source : %ED%2CS%3Cu%FA%CB%F3%5C%16%E0%97q%CF%D1IF%80%85%92