Fixed issue 32153: Update Characteristic process was not working fine
There were 4 request to create datasources for each characteristic, this was happening on one hand because there were two calls to the super and also because  call to OB.Datasource.get() was being executed in two different places. Removing the duplicated super fixed part of the problem, to fix the other part, now a condition has been added in the ob-tree-item class. This condition checks if there is a concrete property (showPopupIconWindow) set to false. If it is set to false the magnifying glass icon will not be created and the OBTreeItemPopupWindow neither, so the uneeded call to the OB.Datasource.get() will not be done.
There was another problem also, if the dropdown of a characteristic was clicked before the creation of the datasource, it was showing the message "No items to show" and a js error in the console. To fix this, a condition has been added. The condition checks if there exists a datasource and clicking the dropdown will not work until a datasource exist.

--HG--
extra : rebase_source : 2331ea3876a8cb2d9440121d89dce4d4b2dc1df9