fixes issue 33622: Selection lost on table/view based P&E having >100 records
Selection could be lost on P&E windows based on a table or a view, having more than 100 records selected. This problem was solved on issue 31989 for those P&E windows based on HQL query tables (this kind of P&E windows make use of a ReadOnlyDataSourceService).

P&E windows based on a table or a view make use of the DefaultDataSourceService which uses DefaultJsonDataService to retrieve data. So, in order to solve the problem, a similar mechanism to the existing one on ReadOnlyDataSourceService has been implemented within DefaultJsonDataService.

Thus, those P&E which use DefaultDataSourceService to retrieve data, now increase the page size dynamically in order to get all selected records in the same request.