fixes issue 33775: Create New is not working with saved views marked as default
When using Create New menu entry to create a new record in a window that has a saved view marked as default, the form was being opened before the loading of the saved view, causing the problem.

To fix it, now the opening of the form is deferred until the window is loaded (including the load of the saved views, if any).