fixes issue 33775: Create New is not working with saved views marked as default
When using Create New menu entry, the opening of the form is deferred until the window is loaded (including the load of the saved views, if any).

Together with this, now the grid data is not loaded when using Create New. This has two main advantages:
  - Better performance: the data will not be loaded when opening the window. If desired, the user can load it after saving/canceling the edition.
  - Homogeneus behavior: this feature will now work as the direct record opening, where the grid is filtered to show just the record being opened.

--HG--
extra : rebase_source : 13bc2cb39cf577b45969be71738781c09648ed63