Fixes issue 33752: Can be closed an order with reserved stock
If the stock was reserved and the associated goods shipment voided, the order was raising an error when trying to close it.
When reservations are retrieved, C_ORDER_POST1 trigger doesn't takes into account the status of previously released reservations.
To make this validation works homogeneously just like is now in other areas where this validation is executed (C_ORDER_POST1, M_INOUT_CREATE, M_INOUT_POST, M_RESERVATION_POST) the query that retrieves the count of reservations of the line also takes into account those lines with status different than 'CL'.
Also validates that is created a new reservation for the line only if doesn't exists any reservation for it.